requirements.txt
Django==3.0.3
djangorestframework==3.11.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
